[TABP-7] Refactor Hotel, City, and Repository Logic

+++ TABP/TABP.API/Contracts/Hotels/CreateHotelRequest.cs
@@ -8,7 +8,7 @@
     byte StarRating,
     double LocationLatitude,
     double LocationLongitude,
-    long CityId,
+    int CityId,
     long OwnerId
     );
 }
